requirements.txt
uvicorn
fastapi~=0.95.2
fastapi-mail

# database
sqlalchemy~=2.0.15
psycopg2-binary
pydantic~=1.10.7
starlette~=0.27.0

# env
python-dotenv~=1.0.0

# token
pyjwt
python-jose
cryptography

# crypto
passlib~=1.7.4
bcrypt
